Fix the PFC view and tests for jinja2 use (#6388)
* Fix the PFC view and tests for jinja2 use

Jinja2 templates run Python code in a more Pythonic way than do Django templates.
Django allows templates to call functions without the standard `()` evocation,
but Jinja allows function execution with `()` and arguments.

The Django test client also does not populate the `context` variable in
Jinja2 responses, so tests can't check `context` for expected values.

In this case, we can just as easily check the `content` of the response
for the values we're looking for.

The technote template was doing a bunch of span replacement actions to pull in
constants that live in the database, but since all these are available to the
view, it makes sense to deliver them straight to the page at render time and
avoid the swapping.

An unrelated nit

The offer view originally allowed a bare response -- rendering the tool without
any school or program specified. Since the tool is useless in this state,
I think it makes sense to return 404 when there is no GET data.

* Update cfgov/paying_for_college/jinja2/paying-for-college/disclosure_technote.html

Thanks.