Order prepaid agreements based on effective_date
Previously, prepaid agreements were ordered based on their created_time,
so the agreement with the latest created_time would be designated "most
recent" in the UI. Now, since the intake process for these agreements
has improved and the effective date field is more reliable, the
stakeholder has asked to use effective_date to order the agreements.

Now, for a prepaid product, the "most recent" agreement will be the one
whose effective date is latest. If multiple agreements share the latest
effective date, the one with the latest created_time will be designated
"most recent".