Refactor filterable lists forms and search
This change refactors our filterable list document search, form, and mixin classes to enable sharing a single instance of the document search class (and its results) in both the mixin class and the form.

This is a first step toward making sure we only do the Elasticsearch query once, and the subsequent `to_queryset` back to Django objects once. With this change we're still making duplicate queries, we're just using the same objects to do it now.