Add new story page type (#6460)
* Add new story page type

Adds a new "Story Page" content type with a sidebar-less centered column
and images that bleed into the left and right gutters. It's use case is
prose-heavy landing pages designed to tell a story.

To let editors choose when they want either RichTextBlock images
or FullWidthText images to bleed into the gutters, a 'bleed' format
option has been added to both image fields. Selecting this option
for an image that's not on a story page will have no affect (it'll
be treated as a standard full width image).

https://docs.wagtail.io/en/stable/advanced_topics/customisation/page_editing_interface.html#rich-text-image-formats

* Properly sort imports to satisfy linter

* Clean up BEM formatting, class names

* Update cfgov/unprocessed/css/enhancements/layout.less

Co-authored-by: Scott Cranfill <scott.cranfill@cfpb.gov>