Update docs for the search_index commands to reflect how we use them. (#6467)
The django-elasticsearch-dsl docs show how you can re-index data for a single model,
but that's not how we use the command in practice, or how we use it in in our Jenkins indexing jobs.

Most of our apps that use Elasticsearch only have one model to index, so specifying a
model is unnecessary. For apps that do index more than one model, it would be rare
to need to re-index only one. The docs should show the most common usage, which is
just specifying an app, as in:

```shell
./cfgov/manage.py search_index --rebuild ask_cfpb
```