Add a tox env for backend test coverage
This small change adds a tox env that will run `diff-cover` on the coverage generated by `unittest-current` or `unittest-future`. This is intended as a convenience means of reporting out the change in test coverage vs `main` that doesn't require running `diff-cover` directly.

I have not changed the GitHub actions for now, as I think it's important to maintain consistency between the backend and frontend actions.