Handle filterable list results without aggregation
PR 6532 added a new Elasticsearch aggregation to filterable list search
results, but neglected to account for cached ES results that may not
have that aggregation.

This commit adds a check so that old cached data won't cause search
pages to 500. The cache will still need to be manually cleared for now
to get the language choices to properly update.

This is a temporary fix that can hopefully be improved with a smarter
strategy overall around these caches (including perhaps getting rid of
them altogether).