Support testing of alternate home page designs
This change adds the ability to use a new NEW_HOME_PAGE feature flag to
toggle between the existing home page designs (both English and Spanish)
to a new (placeholder) design.

This follows the same pattern used previously in PR 5335.

The flag will automatically be enabled on all systems except for
production when ?nhp=True is appended to the URL. So, for example, to
test locally, visit http://localhost:8000/?nhp=True.

The new home page template included here is a copy of the current home
page template, intended to be used as a starting point for development.