Expand try/except around Ask CFPB autocomplete
It seems that Elasticsearch can throw a `RequestError` due to `maxClauseCount ` at any point during the autocomplete query. We had put this `try`/`except` in place as a stop-gap until we have some better user-facing limitations and errors for the Ask CFPB autocomplete (#6537). This expands the `try` to cover both the query initialization and when we access its data.