Add ability to preview the 2021 home page version (#6564)
PR 6560 added a new stub template for a new 2021 version of the website
home page. This commit adds the ability to preview this new version
directly from the Wagtail editor, without having to publish it, and in
the production environment where the new template is not yet enabled.

On the page edit view, the Preview button now becomes a dropdown with
the ability to preview the new template.

This commit also includes a minor edit to the stub template to make it
easier to distinguish from the current one.