Limit Akamai caching of sitemap.xml (#6647)
This commit adds an Edge-Control: no-cache header to the sitemap.xml
view, which changes how Akamai will cache the sitemap.

Currently, the sitemap is cached like most every other page on the site:
served from cache for 7 days without revalidation with the origin, then
revalidated after that and re-sent from origin if necessary. This is
problematic for the sitemap because it should be updated whenever a new
page is published.

With this change, the Akamai-specific Edge-Control header is used,
described thusly in the Akamai docs [0]:

> The Edge-Control header is an Akamai-specific HTTP response header
header that provides controls and parameters on content served through
the edge network. On the Akamai network, Edge-Control settings
settings override Cache-Control and Expires headers.

It works similarly to the standard Cache-Control header, so setting it
to no-cache should work like this [1]:

> The response may be stored by any cache, even if the response is
normally non-cacheable. However, the stored response MUST always go
through validation with the origin server first before using it

If I understand this correctly, this means that requests to sitemap.xml
will always trigger revalidation with the origin, but Akamai will return
304 Not Modified if the content is the same, which should properly use
the Last-Modified header.

To verify this behavior, you can run a local server and do:

$ curl -Is http://localhost:8000/sitemap.xml | grep Edge-Control
Edge-Control: no-cache

[0] https://techdocs.akamai.com/start/docs/caching
[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control