Remove ASK_CFPB_H1 cluster-testing flag
This flag was used in 2018 to facilitate turning a feature on or off
in various "clusters" of pages. The cluster-testing method and the code that enables it
are documented [here](https://cfpb.github.io/consumerfinance.gov/split-testing/).

This change deletes the 2018-specific flag, but leaves open the question of whether
we want to maintain a rather involved custom method for testing clusters of pages.