Simplify the CCDB landing page view and template
The CCDB legacy landing view contained a routine for ensuring that
sample complaint narratives and data were available and up-to-date,
and it handled conditions when different banner messages would be displayed.

This can now be simplified.

Complaint narrative samples are no longer used on the landing page and
don't need to be checked. In addition, the check for stale data never worked,
because of a bug. The code was checking for a nonexistent [response.status](https://github.com/cfpb/consumerfinance.gov/blob/main/cfgov/legacy/views/complaint.py#L48) value
instead of `response.status_code`, and always returned blank json.

Data for the page's high-charts map and metadata for staleness were moved to S3,
where they are updated every morning. Those resources are called from the page.

All this allows us to stop making view calls to Elasticsearch entirely.
The view just needs to deliver the news when the CCDB_TECHNICAL_ISSUES flag is
activated, and that's all we need to test now.

Testing
- Pull in the branch and check that the CCDB landing page renders locally
at http://localhost:8000/data-research/consumer-complaints/
- in the local Wagtail admin, go to /admin/flags/ and enable the flag
` CCDB_TECHNICAL_ISSUES`
- The landing page should now show a warning dialog below the hero:

<img width="1223" alt="ccdb_warning" src="https://user-images.githubusercontent.com/515885/133856030-9ca0bbc1-b8fe-4ca5-a8d7-b51edd10fefa.png">