Fix Lighthouse timeout on pages with video players
Fix to 517307d (PR 6704), which used incorrect logic to prevent an
infinite loop when the video player thumbnail can't load.

We only want to try to load the default thumbnail URL if we haven't
already tried loading it. The previous commit/PR had a bug which failed
to fix the Lighthouse infinite loop problem (described in more detail on
517307d).