Fix translations when previewing Spanish home page
Currently when you preview a Spanish-language HomePage, some of the text
is in English. This is because of some custom preview code that was
added to support the optional previewing of the upcoming 2021 redesign.

This commit alters that logic so the page relies on the
super().serve_preview method, plus a little fix that I've also
contributed upstream to Wagtail core in
https://github.com/wagtail/wagtail/pull/7596.

Fixes internal platform#4255.

To test, preview the Spanish home page, for example:
http://localhost:8000/admin/pages/12786/edit/preview/