Update cached TDP setups when ActivityPages change (#6739)
The ActivitySetUp model stores normally static data used to render the
search form and results for the activity search page. This avoids
having to make dozens of calls to the db and to Elasticsearch to render the
base search page. But when an ActivityPage is changed or added, we need
to update the ActivitySetUp object.

This change adds signal listeners for publishing and unpublishing, and
updates the setups when that happens.

Elasticsearch also needs to be updated when ActivityPages change, but that
is handled by our implementation of django-elasticsearch-dsl.

Testing

Our db has a trashed TDP ActivityPage (11976) that can used locally for testing
by moving it out of the trash.

1. Move the page at http://localhost:8000/admin/pages/11976/ by going to the
MORE -> MOVE command and driving to

```
CFGOV > Consumer Resources > Resources for Practitioners > Youth Financial Education
> Teach the Building Blocks > Search for activities
```

2. Check that the page's PK (in string form) isn't in the TDP setups.

```
./cfgov/manage.py shell
```

```python
from teachers_digital_platform.models import ActivityPage, ActivitySetUp
"11976" in ActivitySetUp.objects.first().card_setup
```

3. Publish the page

```python
page = ActivityPage.objects.get(pk=11976)
page.get_latest_revision().publish()
```

4. Check that the page is now in the setups:

```python
"11976" in ActivitySetUp.objects.first().card_setup
```