Add HTTP Expires header to EventPages if needed
This commit adds the HTTP Expires header to EventPages if they have a
state change in the future, for example before an event begins or during
an event with an end timestamp.

This header will allow for (a) client browsers to properly refetch the
page when its state has changed and (b) downstream caches like Akamai to
automatically expire cached versions of the page.

HTTP Expires headers are of the form "Tue, 04 Jan 2011 00:00:00 GMT".

Addresses internal platform#4081.