Simplify Elasticsearch environment variables (#6748)
* Simplify Elasticsearch environment variables

This commit tries to simplify how we configure Elasticsearch via
environment variables, for virtualenv/local Docker/deployed Docker.

Currently, as of PR 6695, Django will use the combination of ES7_HOST
and ES_PORT to configure the Elasticsearch client. ES_HOST is no longer
used anywhere, since our migration to Elasticsearch 7.x. In Django
settings, these default to localhost/9200 if the environment variables
are not defined.

This PR changes this so we standardize everywhere to ES_HOST and ES_PORT.

With this change, the following configurations are still supported:

1. Local virtualenv - these settings are now no longer set in .env, but
default to localhost/9200 for a locally-running Elasticsearch. The
definitions have been left in .env_SAMPLE commented out for
documentation purposes.

2. Local Docker - port defaults to 9200, but ES_HOST is set in
docker-compose.yml to "elasticsearch" instead.

3. Deployed Docker stack - port defaults to 9200, but ES_HOST is set in
docker-stack.yml to "elasticsearch" instead.

If/when this is merged, there is some cleanup to be done in our internal
Ansible deployment scripts to remove reference to the now-deprecated
ES7_HOST environment variable.