SSO implementation
This change implements support for SAML2 single sign-on based on [djangosaml2](https://djangosaml2.readthedocs.io/) and [pySAML2](https://pysaml2.readthedocs.io/).

It adds environment variables that, when populated, will enable SAML2 authentication, and the required URL patterns, against the SAML2 identity provider with the given metadata URL using the given entity ID for the Django site.

It does require the `xmlsec1` binary to be in place at `/usr/bin/xmlsec1`. That requirement will be handled in our deployment playbooks.