Loosen HOME_PAGE_2021 feature flag conditions
Currently the HOME_PAGE_2021 feature flag conditions mean we only
show the 2021 design if not in production and if the URL query string
includes ?home_page_2021. These are both hardcoded, meaning there is
no way to change them via the Wagtail admin.

For testing purposes, we want to be able to configure the new design
even if that query string isn't specified, but we want to do so via
the Wagtail admin. This change makes that possible.