Apply a new response header, "Edge-Cache-Tag," to the NewsroomLandingPage
The goal is to be able to invalidate Akamai cache on all flavors of
the newsroom landing page – including all variations of query-string
faceting that have been cached.

The standard cache strategy of purging by URL doesn't work in this case,
because there are limitless query string combinations for categories,
topics, languages, date ranges and keywords.

Applying an Edge-Cache-Tag to all instances of the NewsroomLandingPage
allows Akamai to invalidate all the variations it has cached at once.

Cache_tag invalidation is triggered when a NewsroomPage or a LegacyNewsroomPage
gets published.

Testing

We plan to test the triggered cache invalidation on beta.consumerfinance.gov.
In the meantime, you can see that the Edge-Cache-Tag is getting applied to all
variations of /about-us/newsroom/:
- Open your browser dev tools to the Network panel
- Visit <localhost:8000/about-us/newsroom/> and check the response header
for the presence of "Edge-Cache-Tag: newsroom." You may need to hard-refresh
your browser.
- Apply a number of facets and/or search keywords and then see that the result
page also has the tag applied.

This change addresses platform issue #4136