Serve robots.txt from Django
This commit moves robots.txt from a static text file to a Jinja-based
template, and makes it served by Django under the /robots.txt URL.
It also introduces a ROBOTS_TXT_SEARCH_GOV_ONLY feature flag which
modifies robots.txt so that all robots except for "usasearch" are
blocked from the entire site. This flag is enabled on beta, to allow
for a separate beta search index which is only indexed by Search.gov.