Introduce distinct search for beta.consumerfinance.gov (#6788)
* Remove unused show_banner param from header module

The header module show_banner parameter is unused and can be
deprecated. We currently use the BETA_NOTICE feature flag to toggle
the global beta banner on and off.

* Serve robots.txt from Django

This commit moves robots.txt from a static text file to a Jinja-based
template, and makes it served by Django under the /robots.txt URL.
It also introduces a ROBOTS_TXT_SEARCH_GOV_ONLY feature flag which
modifies robots.txt so that all robots except for "usasearch" are
blocked from the entire site. This flag is enabled on beta, to allow
for a separate beta search index which is only indexed by Search.gov.

* Add new search_gov_affiliate Jinja tag

This commit adds a new Jinja tag "search_gov_affiliate" which returns
the appropriate Search.gov affiliate code to use when searching on the
website. The default code is "cfpb", and for Spanish pages we have a
separate Spanish index under "cfpb_es". This commit introduces use of
two new affiliate codes, "cfpb_beta" and "cfpb_beta_es", for use on
beta.consumerfinance.gov.