Stop Akamai from caching the beta_external_testing endpoint
We use the flagged endpoint `/beta_external_testing/` to signal that
testing is in progress on the beta site. When the flag is taken down,
we want the endpoint to immediately return 404, but the cached
200 response is getting cached.

This adds a `Cache-control: no-store` header to the 200 page so that
it is never cached.

This satisfies internal GHE platform issue `#4286`

Testng

- Locally enable the [BETA_EXTERNAL_TESTNIG](http://localhost:8000/admin/flags/BETA_EXTERNAL_TESTING/) flag.
- Open dev tools' network panel and hit <http://localhost:8000/beta_external_testing/>
- Confirm that the endpoint returns a 200 response and that the response headers include "Cache-control: no-store"