Add page path and depth to Elasticsearch index
This commit adds two new fields to the Elasticsearch index that we use
for searching filterable pages: path and depth. These fields exist on
the core Wagtail page model and come from the django-treebeard package
used to organize the page hierarchy.

With these two fields, it is possible to construct a search to that
filters by a wide variety of page relationships: parents, children, etc.
The current Elasticsearch index only includes the page URL, which makes
it harder to, for example, filter only direct children of a given page.

The core Wagtail page filtering based on Django querysets uses page path
and depth, as does the core Wagtail page search.

We have an open internal issue platform#4147 regarding our Events page
(https://www.consumerfinance.gov/about-us/events/) - this page currently
incorrectly includes archived events. This bug was introduced when
filterable page search was migrated to Elasticsearch, because the
current logic doesn't support filtering only to direct children.

This commit only introduces these new fields, and doesn't modify the
search logic to rely on them. That will need to be introduced in a
subsequent PR; I'm splitting this up so that the new index configuration
can get rolled out before any logic that relies on it.

Because this only introduces new fields, there's no real way to test it
beyond inspecting the index manually. One way to do that is to first
reindex pages:

  cfgov/manage.py search_index --rebuild -f --models v1

and then inspect the indexed data, for example in Kibana, to confirm
that the new fields exist.