import_agreements command allows different zip directory structure
Previously, the command would only process a zip file if it contained a
directory structure like /Issuer/agreement.pdf. Now, this command can
process a zip file with an additional directory layer, like
/Containing-folder/Issuer/agreement.pdf. This allows us to use zip files
downloaded directly from Office365.

Previously, the process to update the agreement database included a step
of downloading a file from O365, modifying its directory structure, then
zipping and uploading that file. When that step was done on a Mac, it
introduced encoding errors, so we want to avoid that extra processing if
possible.