Mock less in form_module_handlers tests
This change removes our mocking of a page, a request, and `hasattr` in the tests for the `form_module_handler` context hook. I've kept the mock of `get_streamfields` as it seems a decent way to inject only the field we want to test with into the handler.

This change also removes several superfluous tests that appear to have been built around the mocking.