Update the search index as part of refresh-data.sh
This change adds `./cfgov/manage.py search_index --rebuild -f` to our `refresh-data.sh` script so that we update the search indexes with the contents of the newly refreshed data when we refresh it.

This also moves the command out of the `docker-entrypoint.sh` script that gets run in our Python container. That script includes a call to `refresh-data.sh`.