Add data to the test.sql.gz data dummp for the paying_for_college tests
This turns on GitHub PFC tests and adds a Wagtail page –
/paying-for-college/your-financial-path-to-graduation/ –
to the test data dump, as well as selected entries from these pfc models:
- School
- Alias
- Program
- ConstantCap
- ConstantRate

The school and alias data allow searches and URL-based school selection
to return expected results.

Data for programs and constants are needed because the page makes API
calls to complete forms in the financial-path tool.

After rebuilding the Elasticsearch index with this test data, the cypress
tests succeed.

Testing
- Pull this branch and load your local db with it (locally or in Docker):

```bash
./refresh-data.sh test.sql.gz
```
- Open the cypress test interface:
```
yarn run cypress open --browser chrome
```

Then run each of the 5 tests under `paying-for-college/your-financial-path-to-graduation`