Direct CCDB landing page to Wagtail
We're getting rid of the custom CCDB landing page template and using
a standard Wagtail page instead. This is the first part of that work:
removing the routing that points the CCDB landing page URL to the
custom template and updating functional tests to match. There will be
follow-up work to clean up the old template detritus later.