Navigate to rather than hard coding an Ask answer
This change removes the hard coded Ask answer URL in our funcitonal tests with navigation to the first answer on the Ask landing page.

This hard-coded URL is currently causing our Ask tests to fail in Jenkins because the answer at that URL was unpublished.

This required portal topic tags and a sub landing page to be added to our test data.