Deprecate legacy home page code
We launched the new home page design for English in PR 6919. This PR
removes all legacy English content and Wagtail configuration for the
old English home page design. The new home page content continues to
be stored as static content in an HTML file, not (yet) editable in
Wagtail.

It also migrates the existing Spanish home page content out of
Wagtail and into a static HTML file. We do not anticipate future
editing needs of this legacy design and, if needed, can make changes
by editing the static content.

Moving home page content back into Wagtail for editing will likely be
part of a future effort to add additional languages to the new home
page design.