Reduce model saves in data migrations (#6948)
Our migrate_page_types_and_fields helper function for Wagtail page data
migrations (documented here [0]) takes a "mapper" function that gets
applied to StreamField data to (potentially) alter it. Unfortunately
currently this helper saves every page model instance and revision
that it gets run on, *regardless of whether or not it actually changes
any data*. This results in extra superfluous model saves, which slows
down migrations and could introduce unexpected side effects.

This commit modifies one of our helper functions so that we only save
the page or revision if it has actually made any changes to the data.

Users of our helper function specify the "path" to the block they want
to alter, and the mapper that (potentially) modifies that block data.

Consider for example a nested StreamField block like this, where the
"bar" sub-field has one of a number of choices defined in code.

{
    "type": "myblock",
    "value": {
        "foo": 1234,
        "bar": "choice_a",
        }
    }
}

We want to write a data migration that removes all use of the specific
choice "choice_a" and replace it with "choice_b". But we
don't want to touch pages that have some other value, say, "choice_c".
Under the current approach, those pages (and their revisions) are saved
back to the database, even though their values haven't changed.

This commit attempts to determine whether a migration has actually
modified a StreamField block, and only commit the page back to the
database if so.

[0] https://cfpb.github.io/consumerfinance.gov/migrations/#wagtail-specific-considerations