Add support for hourly salaries in job postings
This change adds the ability to configure JobListingPages with hourly
salaries, as opposed to annual. It does so by adding a new
"salary_is_hourly" boolean field to the page type. If this is selected,
then when the page is rendered it adds "per hour" to the salary line.

In addition, this change also fixes a slight bug where, if salaries
have cents values associated with them (e.g. $15.50/hour), they
would be incorrectly displayed as rounded integers.

To test how this all works, login to the Wagtail admin and visit
http://localhost:8000/admin/template_debug/jobmanager/job_listing_details/