Ensure cache tags exist before invalidating them
This change fixes an issue with our cache tag invalidation on publish of a `AbstractFilterPage` if that page does not belong to a filterable list. In that case, the cache tag invalidation will be attempted but will fail because the list of cache tags to invalidate will be blank.

This fix simply adds a check to ensure the list of cache tags to purge is populated before invalidating.