Fix: Update Elasticsearch when pages are moved
There is currently a bug related to the moving of pages: when a
filterable page is moved within the Wagtail page tree, it is deleted
from its old location but not reindexed at its new location.

This causes problems e.g. when an event page is moved from the list
of future events to the list of past events; it fails to show up in
its new location because it is not being reindexed.

This happens because the django-elasticsearch-dsl index is
configured to index models that are of type AbstractFilterPage [0],
but Wagtail page signals can be called with instances of the base Page
class, for example when a page is moved [1]. We therefore need to
convert pages from the base Page type to the more specific instance
type before trying to reindex them.

[0] https://github.com/cfpb/consumerfinance.gov/blob/bc3e2164b6758e4cca7af76af8985ae05e6ffdd1/cfgov/v1/documents.py#L99
[1] https://github.com/wagtail/wagtail/blob/ed6365ac0ff97505acdc8761ff72cd717f82c149/wagtail/core/models/__init__.py#L1496-L1503