Deprecate Planning for Retirement feedback form
This change removes the feedback form at the bottom of the Planning
for Retirement / Before you Claim page located in English and Spanish
at:

https://www.consumerfinance.gov/consumer-tools/retirement/before-you-claim/
https://www.consumerfinance.gov/consumer-tools/retirement/before-you-claim/es/

At the bottom of the current page there is a form that asks "Was
this page helpful in thinking about when to start collecting benefits?"
with Yes/No buttons. Clicking one of these buttons sends an event to
Google Analytics.

This commit removes this form entirely.

To test, go through the page in both English and Spanish locally:

http://localhost:8000/consumer-tools/retirement/before-you-claim/
http://localhost:8000/consumer-tools/retirement/before-you-claim/es/

Confirm that the page continues to work and there are no browser
console errors.