Add new Research Hub page type (#7023)
* Add research hub filterable page

* Spelling correction in docs: CategoryFilterableMixin

It's not CategoryFilterableListMixin

* Remove outdated test

Now, a page exists (Research Hub Page) that uses the same filterable
page categories that are already used by another page (Research Report
Page). This is intentional -- the Research Hub page is supposed to
filter a subset of Research Reports. So, this test is no longer
enforcing desired behavior. Deleting.