Deprecate feedback forms (#7046)
* Deprecate Wagtail feedback form

This commit deprecates the website feedback form implemented as a
Wagtail StreamField block. It deprecates the block itself, the related
form classes, and related frontend code.

This commit does not delete the Feedback model from the database; this
will be cleaned up in a future PR.

With this change, the conference registration form is the only form
implemented using the AbstractFormBlock approach documented here:
https://cfpb.github.io/consumerfinance.gov/forms-in-wagtail/
The various tests that relied on FeedbackBlock to test this
functionality have been changed to rely on the conference registration
block instead.

Any pages that had a Feedback block in one of their StreamFields
should see that block disappear.

* Deprecate Planning for Retirement feedback form

This change removes the feedback form at the bottom of the Planning
for Retirement / Before you Claim page located in English and Spanish
at:

https://www.consumerfinance.gov/consumer-tools/retirement/before-you-claim/
https://www.consumerfinance.gov/consumer-tools/retirement/before-you-claim/es/

At the bottom of the current page there is a form that asks "Was
this page helpful in thinking about when to start collecting benefits?"
with Yes/No buttons. Clicking one of these buttons sends an event to
Google Analytics.

This commit removes this form entirely.

To test, go through the page in both English and Spanish locally:

http://localhost:8000/consumer-tools/retirement/before-you-claim/
http://localhost:8000/consumer-tools/retirement/before-you-claim/es/

Confirm that the page continues to work and there are no browser
console errors.

* Deprecate custom Buying a House feedback form

Three Buying a House pages use a custom feedback form that asks "Was
this page helpful to you?" and then links to a standalone feedback
page:

- https://www.consumerfinance.gov/owning-a-home/closing-disclosure/
- https://www.consumerfinance.gov/owning-a-home/explore-rates/
- https://www.consumerfinance.gov/owning-a-home/loan-estimate/

This commit deprecates this custom form and all related CSS and JS.

To test, run the frontend build, then run a local server, and check:

- http://localhost:8000/owning-a-home/closing-disclosure/
- http://localhost:8000/owning-a-home/explore-rates/
- http://localhost:8000/owning-a-home/loan-estimate/

* Deprecate Paying for College feedback form

This commit deprecates the standalone feedback form for Paying for
College / Understanding your financial aid offer. This page currently
lives at

https://www.consumerfinance.gov/paying-for-college2/understanding-your-financial-aid-offer/feedback/

and is linked at the bottom of the offer review page.

This commit retires that page and removes the link to it, plus removes
all related frontend assets.

It does not remove the paying_for_college.Feedback model; that will be
cleaned up in the future.

To test with an empty database, add these test fixtures:

cfgov/manage.py loaddata cfgov/paying_for_college/fixtures/test_program.json
cfgov/manage.py loaddata cfgov/paying_for_college/fixtures/test_constants.json

Then visit this URL with a local server:

http://localhost:8000/paying-for-college2/understanding-your-financial-aid-offer/offer/?iped=408039&oid=0000000000000000000000000000000000000000&pid=981

Click through the various prompts and see that the feedback link on the
bottom of the page is now gone.