All-new data model for FIG sections
Previously, the filing instructions guide used the data model that we
had originally designed for the Reports as Webpages project. We borrowed
the code to get the nice design and layout with the Table of Contents
for long, detailed content pages. However, the pages used a structured
clusterable data model instead of a streamfield, and wouldn't allow
wagtail content blocks (e.g. table, well) in the content.
This PR switches the data models to a streamfield while trying to retain
the sidenav TOC functionality.

To do:
 - Re-introduce automatic numbering of content sections, which got
   clobbered by this change
 - Fix sidenav hierarchy
 - Restore JS/CSS for TOC that expands and collapses headers in TOC
   and makes TOC work as accordion on small screens