Fix bug where section IDs weren't calculated for preview
Previously, FIG Content Page used wagtail hooks to assign section IDs to
each section of content only when a user pressed "save" or "publish" in
the Admin. This meant that if the user pressed the "preview" button,
the section assignment would not run, and users would not be able to
preview the section numbering.

Instead, this update overrides the default Wagtail page form to add
custom save functionality, so whenever the user previews, saves, or
publishes the page, section IDs always get assigned, and don't have to
create a new revision.