New models for Filing Instruction Guide content page (#7065)
* All-new data model for FIG sections

Previously, the filing instructions guide used the data model that we
had originally designed for the Reports as Webpages project. We borrowed
the code to get the nice design and layout with the Table of Contents
for long, detailed content pages. However, the pages used a structured
clusterable data model instead of a streamfield, and wouldn't allow
wagtail content blocks (e.g. table, well) in the content.
This PR switches the data models to a streamfield while trying to retain
the sidenav TOC functionality.

To do:
 - Restore JS/CSS for TOC that makes TOC work as accordion on small screens

* Fix sidenav hierarchy: restore nested TOC headers

* Assign section IDs automatically upon page save

* Linting fixes

* Re-order fields in FIG organism, re-make migrations

* Fix bug where section IDs weren't calculated for preview

Previously, FIG Content Page used wagtail hooks to assign section IDs to
each section of content only when a user pressed "save" or "publish" in
the Admin. This meant that if the user pressed the "preview" button,
the section assignment would not run, and users would not be able to
preview the section numbering.

Instead, this update overrides the default Wagtail page form to add
custom save functionality, so whenever the user previews, saves, or
publishes the page, section IDs always get assigned, and don't have to
create a new revision.

* Make get_toc_headers a model method

* Cover case where section is none of the covered types

* More accurate attribute name in context